% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata_file.R
\name{create_metadata}
\alias{create_metadata}
\title{Creates metadata file that can be used to store various metadata inside.}
\usage{
create_metadata(
  ...,
  path = here::here("metadata.csv"),
  overwrite = FALSE,
  backup = TRUE
)
}
\arguments{
\item{...}{Additional parameters. Supports: Column names.}

\item{path}{str | Path to store metadata}

\item{overwrite}{bool | Overwrite old metadata (default = FALSE).}

\item{backup}{bool | Create backup of old metadata (default = TRUE)}
}
\value{
NA
}
\description{
Creates metadata file that can be used to store various metadata inside.
}
\examples{
# Create metadata
create_metadata("constant",
                "column_name",
                "display_name",
                "col_type")
}
