% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metadata.R
\name{rm_metadata}
\alias{rm_metadata}
\title{Removes metadata from metadata fil}
\usage{
rm_metadata(remove, path = here::here("metadata.csv"))
}
\arguments{
\item{remove}{str | Constant you want removed.}

\item{path}{str | Path to metadata (default: 'metadata.csv').}
}
\value{
NA
}
\description{
Removes metadata from metadata fil
}
\examples{
# Create metadata
create_metadata("constant",
                "column_name",
                "display_name",
                "col_type")

# Add metadata
add_metadata(constant = BMI,
             column_name = BMI,
             display_name = "Body mass index (BMI)",
             col_type = "dbl")

# Remove metadata
rm_metadata("BMI")
}
